ekp-ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.4
python-dotenv==1.0.0
httpx==0.26.0

# RAG-Anything 核心依赖
raganything>=0.1.0
lightrag-hku>=0.1.0
openai>=1.0.0

# 本地嵌入模型
sentence-transformers>=2.3.1

# 本地 LLM 支持（通过 Ollama API，无需额外安装）
